Добавил код для рендера по категориям
