+++ src/index.js
@@ -3,7 +3,7 @@ import switcher from './switcher';
 import './js/pop-up-book';
 import './js/fetch-categories';
 import './js/support';
-import './js/renderBooksByCategories';
+import './js/render-books-by-category';
 import './js/categories';
 import './js/top-books';
 import './js/modal-auth/eventsProcessor';
