+++ src/partials/books.html
@@ -1,6 +1,12 @@
-<!-- Это для проверки работоспособности функции renderBooks -->
-
-<div class = "main-container">
-    <h1 class = "main-container-title">Best Sellers Books</h1>
-    <div class = "category-book-container"></div>
+<div class="main-container">
+  <section class="section-categories">
+    <h2 class="section-categories-title">
+      Best Sellers <span class="accent">Books</span>
+    </h2>
+    <ul class="section-categories-list"></ul>
+  </section>
+  <section class="section-books">
+    <h2 class="section-books-title"></h2>
+    <ul class="section-books-list"></ul>
+  </section>
 </div>
